requirements.txt
pyautogui
pytesseract
google-generativeai
python-dotenv
pygetwindow
Pillow
